+++ construct/dagnabitlib/src/test/java/com/onbelay/dagnabit/dagmodel/examples/shortestpath/ModelFixture.java
@@ -6,6 +6,8 @@ import com.onbelay.dagnabit.dagmodel.model.DagRelationship;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.math.BigDecimal;
+
 public class ModelFixture {
 	private static Logger logger = LoggerFactory.getLogger(ModelFixture.class);
 
@@ -23,29 +25,29 @@ public class ModelFixture {
 		DagRelationship link = model.addDefaultRelationship(
 				model.getNode("start"), 
 				model.getNode("A"));
-		link.setWeight(6);
+		link.setWeight(BigDecimal.valueOf(6));
 
 		
 		link = model.addDefaultRelationship(
 				model.getNode("start"), 
 				model.getNode("B"));
-		link.setWeight(2);
+		link.setWeight(BigDecimal.valueOf(2));
 		
 
 		link = model.addDefaultRelationship(
 				model.getNode("B"), 
 				model.getNode("A"));
-		link.setWeight(3);
+		link.setWeight(BigDecimal.valueOf(3));
 		
 		link = model.addDefaultRelationship(
 				model.getNode("A"), 
 				model.getNode("finish"));
-		link.setWeight(1);
+		link.setWeight(BigDecimal.valueOf(1));
 		
 		link = model.addDefaultRelationship(
 				model.getNode("B"), 
 				model.getNode("finish"));
-		link.setWeight(5);
+		link.setWeight(BigDecimal.valueOf(5));
 		
 		
 		return model;
@@ -71,54 +73,54 @@ public class ModelFixture {
 		DagRelationship link = model.addDefaultRelationship(
 				model.getNode("start"), 
 				model.getNode("A"));
-		link.setWeight(6);
+		link.setWeight(BigDecimal.valueOf(6));
 
 		
 		link = model.addDefaultRelationship(
 				model.getNode("start"), 
 				model.getNode("B"));
-		link.setWeight(2);
+		link.setWeight(BigDecimal.valueOf(2));
 		
 		link = model.addDefaultRelationship(
 				model.getNode("start"), 
 				model.getNode("C"));
-		link.setWeight(3);
+		link.setWeight(BigDecimal.valueOf(3));
 		
 
 		link = model.addDefaultRelationship(
 				model.getNode("B"), 
 				model.getNode("A"));
-		link.setWeight(3);
+		link.setWeight(BigDecimal.valueOf(3));
 
 		link = model.addDefaultRelationship(
 				model.getNode("C"), 
 				model.getNode("B"));
-		link.setWeight(1);
+		link.setWeight(BigDecimal.valueOf(1));
 		
 
 		
 		link = model.addDefaultRelationship(
 				model.getNode("A"), 
 				model.getNode("D"));
-		link.setWeight(4);
+		link.setWeight(BigDecimal.valueOf(4));
 
 		
 		link = model.addDefaultRelationship(
 				model.getNode("B"), 
 				model.getNode("D"));
-		link.setWeight(8);
+		link.setWeight(BigDecimal.valueOf(8));
 		
 		
 		link = model.addDefaultRelationship(
 				model.getNode("D"), 
 				model.getNode("finish"));
-		link.setWeight(3);
+		link.setWeight(BigDecimal.valueOf(3));
 		
 		
 		link = model.addDefaultRelationship(
 				model.getNode("C"), 
 				model.getNode("finish"));
-		link.setWeight(13);
+		link.setWeight(BigDecimal.valueOf(13));
 		
 		
 		return model;
